+++ server/server.go
@@ -62,7 +62,7 @@ const (
 	configSwaggerPathFile       			= "swagger_path_file"
 	configShowSwaggerUI         			= "show_swagger_ui"
 	configPoolSize              			= "pool_size"
-)
+	configLocalOfficeURL        			= "docs_local_office_url"
 
 var (
 	logger = ilogger.NewLogger()
@@ -83,6 +83,7 @@ func init() {
 	viper.SetDefault(configSecret, "pandora@drive")
 	viper.SetDefault(configAuthURL, "http://localhost/auth/login")
 	viper.SetDefault(configDocsURL, "http://localhost:3000")
+	viper.SetDefault(configLocalOfficeURL, "http://localhost:3000")
 	viper.SetDefault(configExternalApmURL, "http://localhost:8200")
 	viper.SetDefault(configAllowOrigins, "http://localhost*")
 	viper.SetDefault(configSupportLink, "https://open.rocket.chat")