+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- hotfix:([217](https://github.com/meateam/api-gateway/pull/217)): added authType and appID to metrics.
+- hotfix:([223](https://github.com/meateam/api-gateway/pull/223)): added authType and appID to metrics.
 ## [v3.1.0] - 2021-03-14
 
 ### Added