+++ CHANGELOG.md
@@ -8,9 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [v3.1.0] - 2021-02-22
 ### Added
-- FEAT([218](https://github.com/meateam/api-gateway/pull/218)): add advance search by mail/T.
+- major: ([218](https://github.com/meateam/api-gateway/pull/218)): add advance search by mail/T.
 
 ## [v3.0.0] - 2021-01-13